# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Green Way Motors API
  version: 0.3.0
  description: Implemented backend contract. GreenWay_Backend_Master_Spec_v1.md is the product and API roadmap.
servers:
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /api/v1/vehicles:
    get:
      tags: [Vehicles]
      operationId: listPublishedVehicles
      parameters:
        - {
            name: market,
            in: query,
            required: true,
            description: Active destination market slug.,
            schema: { type: string }
          }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100 }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
        - { name: makeId, in: query, schema: { type: integer } }
        - { name: modelId, in: query, schema: { type: integer } }
        - { name: bodyTypeId, in: query, schema: { type: integer } }
        - { name: condition, in: query, schema: { type: string, enum: [new, used] } }
        - { name: yearFrom, in: query, schema: { type: integer } }
        - { name: yearTo, in: query, schema: { type: integer } }
        - { name: mileageFrom, in: query, schema: { type: integer } }
        - { name: mileageTo, in: query, schema: { type: integer } }
        - { name: engineCcFrom, in: query, schema: { type: integer } }
        - { name: engineCcTo, in: query, schema: { type: integer } }
        - {
            name: fuel,
            in: query,
            description: Comma-separated or repeated enum values.,
            schema: { type: string }
          }
        - { name: transmission, in: query, schema: { type: string } }
        - { name: drive, in: query, schema: { type: string } }
        - { name: steering, in: query, schema: { type: string } }
        - { name: stockCountryId, in: query, schema: { type: integer } }
        - {
            name: featureIds,
            in: query,
            description: Comma-separated feature IDs.,
            schema: { type: string }
          }
        - {
            name: sort,
            in: query,
            schema:
              {
                type: string,
                enum: [newest, oldest, year_desc, year_asc, mileage_asc, mileage_desc]
              }
          }
      responses:
        '200': { $ref: '#/components/responses/PublicVehicles' }
        '422': { $ref: '#/components/responses/Problem' }
  /api/v1/vehicles/{referenceNo}:
    get:
      tags: [Vehicles]
      operationId: getPublishedVehicle
      parameters:
        - {
            name: market,
            in: query,
            required: true,
            description: Active destination market slug.,
            schema: { type: string }
          }
        - {
            name: referenceNo,
            in: path,
            required: true,
            schema: { type: string, pattern: '^GW-[0-9]+$' }
          }
      responses:
        '200': { $ref: '#/components/responses/PublicVehicle' }
        '404': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
...
  /api/v1/admin/vehicles:
    get:
      tags: [Admin - Vehicles]
      operationId: adminListVehicles
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
        - { name: q, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
        - { name: availabilityStatus, in: query, schema: { type: string } }
      responses: { '200': { $ref: '#/components/responses/Page' } }
    post:
      tags: [Admin - Vehicles]
      operationId: adminCreateVehicle
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/VehicleInput' } } }
      responses:
        '201': { $ref: '#/components/responses/Data' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/vehicles/{id}:
    parameters: [{ $ref: '#/components/parameters/UuidId' }]
    get:
      tags: [Admin - Vehicles]
      operationId: adminGetVehicle
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Data' }
        '404': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Admin - Vehicles]
      operationId: adminUpdateVehicle
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/VehiclePatch' } } }
      responses:
        '200': { $ref: '#/components/responses/Data' }
        '422': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [Admin - Vehicles]
      operationId: adminSoftDeleteVehicle
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Vehicle archived and soft deleted. }
        '409': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/vehicles/{id}/restore:
    parameters: [{ $ref: '#/components/parameters/UuidId' }]
    post:
      tags: [Admin - Vehicles]
      operationId: adminRestoreVehicle
      security: [{ bearerAuth: [] }]
      responses: { '200': { $ref: '#/components/responses/Data' } }
...
  /api/v1/admin/vehicles/{id}/features:
    parameters: [{ $ref: '#/components/parameters/UuidId' }]
    get:
      tags: [Admin - Vehicles]
      operationId: adminGetVehicleFeatures
      security: [{ bearerAuth: [] }]
      responses: { '200': { $ref: '#/components/responses/Data' } }
    put:
      tags: [Admin - Vehicles]
      operationId: adminReplaceVehicleFeatures
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/FeatureIdsInput' } } }
      responses:
        '200': { $ref: '#/components/responses/Data' }
        '422': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/vehicles/{id}/markets:
    parameters: [{ $ref: '#/components/parameters/UuidId' }]
    get:
      tags: [Admin - Markets]
      operationId: adminGetVehicleMarkets
      security: [{ bearerAuth: [] }]
      responses: { '200': { $ref: '#/components/responses/Data' } }
    put:
      tags: [Admin - Markets]
      operationId: adminReplaceVehicleMarkets
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/VehicleMarketsInput' } } }
      responses:
        '200': { $ref: '#/components/responses/Data' }
        '422': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/vehicles/{id}/costing:
    parameters: [{ $ref: '#/components/parameters/UuidId' }]
    get:
      tags: [Admin - Costing]
      operationId: adminGetVehicleCosting
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Data' }
        '403': { $ref: '#/components/responses/Problem' }
    put:
      tags: [Admin - Costing]
      operationId: adminReplaceVehicleCosting
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/VehicleCostingInput' } } }
      responses:
        '200': { $ref: '#/components/responses/Data' }
        '403': { $ref: '#/components/responses/Problem' }
...
components:
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: greenway_refresh_token
  responses:
    PublicVehicle:
      description: Published, available vehicle eligible for the selected destination market. Source and cost fields are absent.
      content:
        application/json:
          schema:
            type: object
            required: [data, meta]
            properties:
              data: { $ref: '#/components/schemas/PublicVehicle' }
              meta: { $ref: '#/components/schemas/RequestMeta' }
    PublicVehicles:
      description: Market-eligible published and available vehicles.
      content:
        application/json:
          schema:
            type: object
            required: [data, meta]
            properties:
              data: { type: array, items: { $ref: '#/components/schemas/PublicVehicle' } }
              meta: { type: object, additionalProperties: true }
...
    Data:
      description: Permission-scoped data with request ID metadata.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/DataEnvelope' }
    Page:
      description: Cursor-paginated permission-scoped collection.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/PageEnvelope' }
    Problem:
      description: RFC 9457 problem details.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    RequestMeta:
      type: object
      required: [requestId]
      properties: { requestId: { type: string } }
...
    VehicleInput:
      type: object
      additionalProperties: false
      required: [makeId, modelId, stockCountryId, title, year]
      properties:
        makeId: { type: integer, minimum: 1 }
        modelId: { type: integer, minimum: 1 }
        stockCountryId: { type: integer, minimum: 1 }
        title: { type: string, minLength: 1 }
        year: { type: integer, minimum: 1950, maximum: 2100 }
        inventorySourceId: { type: [string, 'null'], format: uuid }
        stockNumber: { type: [string, 'null'] }
        vin: { type: [string, 'null'] }
        condition: { type: string, enum: [new, used] }
        variant: { type: [string, 'null'] }
        bodyTypeId: { type: [integer, 'null'] }
        mileageKm: { type: [integer, 'null'] }
        engineCc: { type: [integer, 'null'] }
        fuel: { type: [string, 'null'] }
        transmission: { type: [string, 'null'] }
        drive: { type: [string, 'null'] }
        steering: { type: [string, 'null'] }
        seats: { type: [integer, 'null'] }
        doors: { type: [integer, 'null'] }
        exteriorColor: { type: [string, 'null'] }
        interiorColor: { type: [string, 'null'] }
        stockCity: { type: [string, 'null'] }
        description: { type: [string, 'null'] }
    VehiclePatch:
      type: object
      additionalProperties: false
      properties:
        makeId: { type: integer }
        modelId: { type: integer }
        stockCountryId: { type: integer }
        title: { type: string }
        year: { type: integer }
        inventorySourceId: { type: [string, 'null'], format: uuid }
        stockNumber: { type: [string, 'null'] }
        vin: { type: [string, 'null'] }
        condition: { type: string, enum: [new, used] }
        variant: { type: [string, 'null'] }
        bodyTypeId: { type: [integer, 'null'] }
        mileageKm: { type: [integer, 'null'] }
        engineCc: { type: [integer, 'null'] }
        fuel: { type: [string, 'null'] }
        transmission: { type: [string, 'null'] }
        drive: { type: [string, 'null'] }
        steering: { type: [string, 'null'] }
        seats: { type: [integer, 'null'] }
        doors: { type: [integer, 'null'] }
        exteriorColor: { type: [string, 'null'] }
        interiorColor: { type: [string, 'null'] }
        stockCity: { type: [string, 'null'] }
        description: { type: [string, 'null'] }
    FeatureIdsInput:
      type: object
      additionalProperties: false
      required: [featureIds]
      properties:
        { featureIds: { type: array, uniqueItems: true, items: { type: integer, minimum: 1 } } }
...